+++ lib/td_ai/prompt_parser.ex
@@ -32,10 +32,16 @@ defmodule TdAi.PromptParser do
     end)
   end
 
-  def generate_user_prompt(%Prompt{user_prompt_template: user_prompt_template}, fields, resource) do
+  def generate_user_prompt(
+        %Prompt{user_prompt_template: user_prompt_template},
+        fields,
+        resource,
+        opts \\ []
+      ) do
     user_prompt_template
     |> String.replace("{fields}", Jason.encode!(fields))
     |> String.replace("{resource}", Jason.encode!(resource))
+    |> String.replace("{locales}", Jason.encode!(Keyword.get(opts, :locales)))
   end
 
   defp put_value(_, _, :error), do: :error